+++ src/styles/styles.css
@@ -20,7 +20,7 @@
   position: relative;
   /* overflow-x: hidden; */
 }
-.color-filter{
+.color-filter {
   width: 100vw;
   height: 100vh;
   background: linear-gradient(var(--dark-purple), var(--royal-purple));
@@ -30,46 +30,6 @@
   top: 0;
   left: 0;
   z-index: -3;
-}
-.navBar {
-  width: 100%;
-  color: var(--aquamarine);
-  align-items: center;
-  font-size: 1.2rem;
-  background: linear-gradient(
-    rgba(125, 125, 121, 0),
-    rgba(147, 177, 174, 0.1),
-    rgba(152, 148, 189, 0.548)
-  );
-  top: 0;
-  padding: 0 20%;
-  display: flex;
-  justify-content: space-between;
-  height: 5rem;
-  /* border-bottom: 1px solid rgba(113, 255, 208, 1); */
-}
-.navBar__link {
-  color: var(--aquamarine);
-  padding: 0.5rem 0;
-  line-height: 1.5ch;
-  border-bottom: 1px solid transparent;
-  border-top: 1px solid transparent;
-  transition: 0.5s;
-}
-.navBar__link:hover {
-  border-bottom: 1px solid var(--aquamarine);
-  border-top: 1px solid var(--aquamarine);
-  text-shadow: 1px 1px 5px var(--aquamarine);
-  padding: 1rem 0;
-  transition: 0.5s;
-}
-.navBar__RickAndMorty {
-  background: url(../assets/images/trippy.jpg);
-  /* background-size: cover; */
-  /* border-bottom: 1px solid var(--aquamarine); */
-  /* box-shadow: 0 0 150px 5px rgba(12, 255, 174, 0.3); */
-  position: fixed;
-  z-index: 2;
 }
 
 .bg-img {
